+++ src/Extensions/ReadOnlySpan/Linq/Take.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace SpanExtensions
+{
+    public static partial class ReadOnlySpanExtensions
+    {
+        /// <summary>
+        /// Returns a specified number of contiguous elements from the start of a <see cref="ReadOnlySpan{T}"/>.
+        /// </summary>
+        /// <typeparam name="T">The type of elements in the <see cref="ReadOnlySpan{T}"/>.</typeparam>
+        /// <param name="source">The <see cref="ReadOnlySpan{T}"/> to operate on.</param>
+        /// <param name="count">The Number of elements to take.</param>
+        /// <returns>A <see cref="ReadOnlySpan{T}"/> that contains <paramref name="count"/> elements from the start of <paramref name="source"/>.</returns>
+        /// <exception cref="ArgumentNullException"><paramref name="source"/> is null.</exception>
+        public static ReadOnlySpan<T> Take<T>(this ReadOnlySpan<T> source, int count)
+        {
+            return source[..count];
+        }
+    }
+}
