+++ src/Extensions/ReadOnlySpan/String/Remove.cs
